+++ projeto/lib/app/layers/presenters/view/details/widgets/info_person_row_widget.dart
@@ -1,0 +1,28 @@
+import 'package:flutter/material.dart';
+
+class InfoPersonRow extends StatelessWidget {
+  const InfoPersonRow({
+    Key? key,
+    required this.title,
+    required this.value,
+  }) : super(key: key);
+
+  final String title;
+  final String value;
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.only(bottom: 8.0),
+      child: Row(
+        children: [
+          Text(title),
+          Text(
+            value,
+            style: Theme.of(context).textTheme.bodyText1,
+          ),
+        ],
+      ),
+    );
+  }
+}
